Fix re-authentication not working after logout
It is not possible to re-authenticate a user with
a previous refresh token without stopping/starting AACS.

To avoid going trough the whole CBL process again we might
save the refresh token to speed up the authorization process.

However, after logging out and starting the authentication
process again the CBLAuthorizationProvider gets stuck on
the REFRESHING_TOKEN state.

This is because we don't reset the m_timeToRefresh variable
after stopping a previous authorization process.
That way the new authorization will be blocked until
that time expires.